- don't follow symlinks using chown/chgrp from --setugids pipe.    - fix PGP clearsign signature verification by correctly skipping RFC2440 "dash-escapes"    - fix line-continuation handling in macro definition parsing    - permit opt-in network access using %_rpmgio macro.    - fix: deny rpmioSlurp network access (like all other Fopen call's).    - rpm.org: Assorted int -> size_t corrections for string lengths.    - rpm.org: More assorted int -> size_t uses.

+++ rpmio/stringbuf.h
@@ -49,7 +49,7 @@ void stripTrailingBlanksStringBuf(StringBuf sb)
 
 /**
  */
-void appendStringBufAux(StringBuf sb, const char * s, int nl)
+void appendStringBufAux(StringBuf sb, const char * s, size_t nl)
 	/*@modifies sb @*/;
 
 #ifdef __cplusplus
